requirements.txt
streamlit
plotly
groq
python-dotenv
